Use API in the veBAL generator (#4199)
* Use API in the veBAL generator
